Move SharedViewModel to Utils package
