+++ app/src/main/java/ch/heigvd/pro/b04/android/Poll/PollAdapter.java
@@ -16,6 +16,7 @@ import java.util.List;
 
 import ch.heigvd.pro.b04.android.Datamodel.Question;
 import ch.heigvd.pro.b04.android.R;
+import ch.heigvd.pro.b04.android.Utils.SharedViewModel;
 
 public class PollAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder>  {
     private static final int VIEW_TYPE_HEADER = 0;
