+++ app/src/main/java/ch/heigvd/pro/b04/android/Poll/PollViewModel.java
@@ -7,6 +7,7 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
 import ch.heigvd.pro.b04.android.Datamodel.Question;
+import ch.heigvd.pro.b04.android.Utils.SharedViewModel;
 
 public class PollViewModel extends SharedViewModel {
     private MutableLiveData<Question> questionToView = new MutableLiveData<>();
