+++ app/src/main/java/ch/heigvd/pro/b04/android/Poll/PollViewModelFactory.java
@@ -6,6 +6,8 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
+import ch.heigvd.pro.b04.android.Utils.SharedViewModel;
+
 /**
  * An implementation of a {@link ViewModelProvider.Factory} that supports the creation of some
  * {@link SharedViewModel}s with specific authentication tokens, and dedicated poll identifiers as
